feat(tree-builder): add update_tree method to synchronize directory tree state with file system
The update_tree method allows dynamically reflecting changes to the directory structure, including additions, deletions, and modifications.

BREAKING CHANGE: Updated the FileNode 'modified' attribute from float to int to align with nanosecond precision timestamps.

test(update-tree): add test cases for update_tree functionality in test_update_tree.py

+++ pyarbor/node.py
@@ -27,7 +27,7 @@ class FileNode(Node):
     """Represents a file node in the tree structure."""
 
     type: str = "file"
-    modified: float = Field(..., title="Last modified timestamp of the file")
+    modified: int = Field(..., title="Last modified timestamp of the file")
     content: str | Tree | None = Field(
         ..., title="Content of the file as a string (not parsed) or Tree-sitter Tree"
     )
